# simd: AVX256 int8 dot product