--- a/adapters/repos/db/vector/hnsw/distancer/asm/dot_amd64.go
+++ b/adapters/repos/db/vector/hnsw/distancer/asm/dot_amd64.go
@@ -14,7 +14,9 @@ package asm
 //go:generate goat ../c/dot_avx256_amd64.c -O3 -mavx2 -mfma -mavx512f -mavx512dq -e="-mfloat-abi=hard" -e="-Rpass-analysis=loop-vectorize" -e="-Rpass=loop-vectorize" -e="-Rpass-missed=loop-vectorize"
 //go:generate goat ../c/dot_avx512_amd64.c -O3 -mavx2 -mfma -mavx512f -mavx512dq -e="-mfloat-abi=hard" -e="-Rpass-analysis=loop-vectorize" -e="-Rpass=loop-vectorize" -e="-Rpass-missed=loop-vectorize"
 
-import "unsafe"
+import (
+	"unsafe"
+)
 
 func DotAVX256(x []float32, y []float32) float32 {
 	var res float32
@@ -34,6 +36,19 @@ func DotAVX512(x []float32, y []float32) float32 {
 
 	l := len(x)
 	dot_512(
+		unsafe.Pointer(unsafe.SliceData(x)),
+		unsafe.Pointer(unsafe.SliceData(y)),
+		unsafe.Pointer(&res),
+		unsafe.Pointer(&l))
+
+	return res
+}
+
+func DotByteAVX256(x []uint8, y []uint8) uint32 {
+	var res uint32
+
+	l := len(x)
+	dot_byte_256(
 		unsafe.Pointer(unsafe.SliceData(x)),
 		unsafe.Pointer(unsafe.SliceData(y)),
 		unsafe.Pointer(&res),