--- a/adapters/repos/db/vector/hnsw/distancer/asm/dot_amd64.go
+++ b/adapters/repos/db/vector/hnsw/distancer/asm/dot_amd64.go
@@ -45,6 +45,33 @@ func DotAVX512(x []float32, y []float32) float32 {
 }
 
 func DotByteAVX256(x []uint8, y []uint8) uint32 {
+	switch len(x) {
+	case 2:
+		return dot2[uint8, uint32](x, y)
+	case 3:
+		return dot3[uint8, uint32](x, y)
+	case 4:
+		return dot4[uint8, uint32](x, y)
+	case 5:
+		return dot5[uint8, uint32](x, y)
+	case 6:
+		return dot6[uint8, uint32](x, y)
+	case 7:
+		return dot7[uint8, uint32](x, y)
+	case 8:
+		// manually inlined dot8(x, y)
+		sum := x[7]*y[7] + x[6]*y[6]
+		return dot6[uint8, uint32](x, y) + uint32(sum)
+	case 10:
+		// manually inlined dot10(x, y)
+		sum := x[9]*y[9] + x[8]*y[8] + x[7]*y[7] + x[6]*y[6]
+		return dot6[uint8, uint32](x, y) + uint32(sum)
+	case 12:
+		// manually inlined dot12(x, y)
+		sum := x[11]*y[11] + x[10]*y[10] + x[9]*y[9] + x[8]*y[8] + x[7]*y[7] + x[6]*y[6]
+		return dot6[uint8, uint32](x, y) + uint32(sum)
+	}
+
 	var res uint32
 
 	l := len(x)