--- a/adapters/repos/db/vector/hnsw/distancer/asm/dot_inline.go
+++ b/adapters/repos/db/vector/hnsw/distancer/asm/dot_inline.go
@@ -18,38 +18,48 @@ package asm
 // - it allows dot2, dot4 and dot6 to be inlined (the other ones are too large)
 // See go tool compile -d=ssa/check_bce/debug=1 -m dot_inline.go
 
-func dot2(x []float32, y []float32) float32 {
+type number interface {
+	~uint8 | ~uint32 | ~float32
+}
+
+//gcassert:inline
+func dot2[T number, U number](x []T, y []T) U {
 	sum := x[1]*y[1] + x[0]*y[0]
 
-	return sum
+	return U(sum)
 }
 
-func dot4(x []float32, y []float32) float32 {
+//gcassert:inline
+func dot3[T, U number](x []T, y []T) U {
+	sum := x[2] * y[2]
+
+	return dot2[T, U](x, y) + U(sum)
+}
+
+//gcassert:inline
+func dot4[T, U number](x []T, y []T) U {
 	sum := x[3]*y[3] + x[2]*y[2]
 
-	return dot2(x, y) + sum
+	return dot2[T, U](x, y) + U(sum)
 }
 
-func dot6(x []float32, y []float32) float32 {
+//gcassert:inline
+func dot5[T, U number](x []T, y []T) U {
+	sum := x[4] * y[4]
+
+	return dot4[T, U](x, y) + U(sum)
+}
+
+//gcassert:inline
+func dot6[T, U number](x []T, y []T) U {
 	sum := x[5]*y[5] + x[4]*y[4]
 
-	return dot4(x, y) + sum
+	return dot4[T, U](x, y) + U(sum)
 }
 
-func dot8(x []float32, y []float32) float32 {
-	sum := x[7]*y[7] + x[6]*y[6]
+//gcassert:inline
+func dot7[T, U number](x []T, y []T) U {
+	sum := x[6] * y[6]
 
-	return dot6(x, y) + sum
-}
-
-func dot10(x []float32, y []float32) float32 {
-	sum := x[9]*y[9] + x[8]*y[8] + x[7]*y[7] + x[6]*y[6]
-
-	return dot6(x, y) + sum
-}
-
-func dot12(x []float32, y []float32) float32 {
-	sum := x[11]*y[11] + x[10]*y[10] + x[9]*y[9] + x[8]*y[8] + x[7]*y[7] + x[6]*y[6]
-
-	return dot6(x, y) + sum
+	return dot6[T, U](x, y) + U(sum)
 }