--- a/adapters/repos/db/vector/hnsw/distancer/dot_product.go
+++ b/adapters/repos/db/vector/hnsw/distancer/dot_product.go
@@ -49,15 +49,6 @@ func NewDotProductProvider() DotProductProvider {
 	return DotProductProvider{}
 }
 
-func DotProductGo(a, b []float32) float32 {
-	var sum float32
-	for i := range a {
-		sum += a[i] * b[i]
-	}
-
-	return -sum
-}
-
 func (d DotProductProvider) SingleDist(a, b []float32) (float32, bool, error) {
 	if len(a) != len(b) {
 		return 0, false, errors.Errorf("vector lengths don't match: %d vs %d",