--- a/adapters/repos/db/vector/hnsw/distancer/dot_product_amd64_test.go
+++ b/adapters/repos/db/vector/hnsw/distancer/dot_product_amd64_test.go
@@ -18,8 +18,28 @@ import (
 	"unsafe"
 
 	"github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer/asm"
+	"golang.org/x/exp/constraints"
 	"golang.org/x/sys/cpu"
 )
+
+func DotProductGo[T constraints.Integer | constraints.Float](a, b []T) T {
+	var sum T
+	for i := range a {
+		sum += a[i] * b[i]
+	}
+
+	return sum
+}
+
+var dotByteImpl func(a, b []byte) uint32 = func(a, b []byte) uint32 {
+	var sum uint16
+
+	for i := range a {
+		sum += uint16(a[i]) * uint16(b[i])
+	}
+
+	return uint32(sum)
+}
 
 func testDotProductFixedValue(t *testing.T, size uint, dotFn func(x []float32, y []float32) float32) {
 	count := 10000
@@ -38,7 +58,7 @@ func testDotProductFixedValue(t *testing.T, size uint, dotFn func(x []float32, y
 			panic("NaN")
 		}
 
-		resControl := DotProductGo(vec1, vec2)
+		resControl := -DotProductGo(vec1, vec2)
 		delta := float64(0.01)
 		diff := float64(resControl) - float64(res)
 		if diff < -delta || diff > delta {
@@ -78,7 +98,7 @@ func testDotProductRandomValue(t *testing.T, size uint, dotFn func(x []float32, 
 			panic("NaN")
 		}
 
-		resControl := DotProductGo(vec1s[i], vec2s[i])
+		resControl := -DotProductGo(vec1s[i], vec2s[i])
 		delta := float64(0.01)
 		diff := float64(resControl) - float64(res)
 		if diff < -delta || diff > delta {
@@ -134,6 +154,119 @@ func TestCompareDotProductImplementations(t *testing.T) {
 				testDotProductFixedValue(t, size, asm.DotAVX512)
 				testDotProductRandomValue(t, size, asm.DotAVX512)
 			}
+		})
+	}
+}
+
+func testDotProductByteFixedValue(t *testing.T, size uint, dotFn func(x []uint8, y []uint8) uint32) {
+	vec1 := make([]uint8, size)
+	vec2 := make([]uint8, size)
+	for i := range vec1 {
+		vec1[i] = 1
+		vec2[i] = 1
+	}
+	res := dotFn(vec1, vec2)
+
+	resControl := dotByteImpl(vec1, vec2)
+	if uint32(resControl) != res {
+		t.Logf("for dim: %d -> want: %d, got: %d", size, resControl, res)
+		t.Fail()
+	}
+}
+
+func testDotProductByteRandomValue(t *testing.T, size uint, dotFn func(x []byte, y []byte) uint32) {
+	r := getRandomSeed()
+	count := 10000
+
+	vec1s := make([][]byte, count)
+	vec2s := make([][]byte, count)
+
+	for i := 0; i < count; i++ {
+		vec1 := make([]byte, size)
+		vec2 := make([]byte, size)
+		for j := range vec1 {
+			vec1[j] = byte(r.Uint32() % 256)
+			vec2[j] = byte(r.Uint32() % 256)
+		}
+	}
+
+	for i := 0; i < count; i++ {
+		res := dotFn(vec1s[i], vec2s[i])
+
+		resControl := dotByteImpl(vec1s[i], vec2s[i])
+		if uint32(resControl) != res {
+			t.Logf("for dim: %d -> want: %d, got: %d", size, resControl, res)
+			t.Fail()
+		}
+	}
+}
+
+func TestCompareDotProductByte(t *testing.T) {
+	sizes := []uint{
+		1,
+		2,
+		3,
+		4,
+		5,
+		6,
+		8,
+		10,
+		12,
+		16,
+		24,
+		30,
+		31,
+		32,
+		64,
+		67,
+		128,
+		256,
+		260,
+		299,
+		300,
+		384,
+		390,
+		600,
+		768,
+		777,
+		784,
+		1024,
+		1536,
+	}
+
+	for _, size := range sizes {
+		t.Run(fmt.Sprintf("with size %d", size), func(t *testing.T) {
+			testDotProductByteFixedValue(t, size, asm.DotByteAVX256)
+			testDotProductByteRandomValue(t, size, asm.DotByteAVX256)
+		})
+	}
+}
+
+func benchmarkDotByte(b *testing.B, dims int, dotFn func(a, b []byte) uint32) {
+	r := getRandomSeed()
+
+	vec1 := make([]byte, dims)
+	vec2 := make([]byte, dims)
+	for i := range vec1 {
+		vec1[i] = byte(r.Uint32() % 256)
+		vec2[i] = byte(r.Uint32() % 256)
+	}
+
+	b.ResetTimer()
+	for n := 0; n < b.N; n++ {
+		dotFn(vec1, vec2)
+	}
+}
+
+func BenchmarkDotByte(b *testing.B) {
+	dims := []int{2, 4, 6, 8, 10, 12, 16, 24, 30, 32, 128, 256, 300, 384, 512, 768, 1024, 1536}
+	for _, dim := range dims {
+		b.Run(fmt.Sprintf("%d dimensions", dim), func(b *testing.B) {
+			// benchmarkDotByte(b, dim, dotByteImpl)
+			benchmarkDotByte(b, dim, asm.DotByteAVX256)
+
+			// b.Run("pure go", func(b *testing.B) { benchmarkDotByte(b, dim, dotByteImpl) })
+			// b.Run("avx", func(b *testing.B) { benchmarkDotByte(b, dim, asm.DotByteAVX256) })
 		})
 	}
 }