### What's being changed:

This adds an int8 dot product implementation using AVX256.

```
goos: linux
goarch: amd64
pkg: github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer
cpu: Intel(R) Core(TM) i7-8700K CPU @ 3.70GHz
                          │   1_go.txt   │        7_avx_256_inline.txt         │
                          │    sec/op    │    sec/op     vs base               │
DotByte/2_dimensions-4       3.001n ± 1%   2.282n ±  4%  -23.95% (p=0.002 n=6)
DotByte/4_dimensions-4       4.277n ± 3%   2.797n ±  2%  -34.60% (p=0.002 n=6)
DotByte/6_dimensions-4       3.976n ± 9%   4.397n ±  1%  +10.57% (p=0.002 n=6)
DotByte/8_dimensions-4       4.936n ± 1%   5.351n ± 11%   +8.39% (p=0.002 n=6)
DotByte/10_dimensions-4      5.377n ± 1%   5.768n ±  2%   +7.28% (p=0.002 n=6)
DotByte/12_dimensions-4      6.209n ± 9%   6.202n ±  3%        ~ (p=0.461 n=6)
DotByte/16_dimensions-4      7.726n ± 1%   6.366n ±  3%  -17.61% (p=0.002 n=6)
DotByte/24_dimensions-4     11.070n ± 2%   9.436n ±  0%  -14.76% (p=0.002 n=6)
DotByte/30_dimensions-4      13.61n ± 8%   11.97n ± 20%  -12.02% (p=0.039 n=6)
DotByte/32_dimensions-4     14.395n ± 1%   5.131n ±  2%  -64.36% (p=0.002 n=6)
DotByte/128_dimensions-4    62.455n ± 0%   8.582n ±  0%  -86.26% (p=0.002 n=6)
DotByte/256_dimensions-4    118.30n ± 8%   12.49n ±  8%  -89.45% (p=0.002 n=6)
DotByte/300_dimensions-4    137.45n ± 0%   20.30n ±  1%  -85.23% (p=0.002 n=6)
DotByte/384_dimensions-4    174.05n ± 1%   16.18n ±  1%  -90.71% (p=0.002 n=6)
DotByte/512_dimensions-4    230.80n ± 0%   19.77n ±  7%  -91.44% (p=0.002 n=6)
DotByte/768_dimensions-4    342.50n ± 7%   26.78n ±  1%  -92.18% (p=0.002 n=6)
DotByte/1024_dimensions-4   453.05n ± 0%   33.95n ±  0%  -92.51% (p=0.002 n=6)
DotByte/1536_dimensions-4   677.55n ± 6%   48.80n ±  2%  -92.80% (p=0.002 n=6)
geomean                      30.75n        9.730n        -68.36%
```

### Review checklist

- [ ] Documentation has been updated, if necessary. Link to changed documentation:
- [ ] Chaos pipeline run or not necessary. Link to pipeline:
- [x] All new code is covered by tests where it is reasonable.
- [x] Performance tests have been run or not necessary.